Split reactor API models into sensor and trigger models
