+++ st2api/st2api/controllers/v1/sensors.py
@@ -16,7 +16,7 @@
 import six
 from st2common import log as logging
 from st2common.persistence.reactor import SensorType
-from st2common.models.api.reactor import SensorTypeAPI
+from st2common.models.api.sensor import SensorTypeAPI
 from st2api.controllers import resource
 
 http_client = six.moves.http_client
