+++ st2api/st2api/controllers/v1/triggers.py
@@ -19,7 +19,7 @@ from pecan.rest import RestController
 import six
 
 from st2common import log as logging
-from st2common.models.api.reactor import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
+from st2common.models.api.trigger import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
 from st2common.models.api.base import jsexpose
 from st2common.models.system.common import ResourceReference
 from st2common.persistence.reactor import TriggerType, Trigger, TriggerInstance
