+++ st2common/st2common/models/api/rule.py
@@ -18,7 +18,7 @@ import copy
 import six
 
 from st2common.models.api.base import BaseAPI
-from st2common.models.api.reactor import TriggerAPI
+from st2common.models.api.trigger import TriggerAPI
 from st2common.models.api.tag import TagsHelper
 from st2common.models.db.reactor import RuleDB, ActionExecutionSpecDB
 from st2common.models.system.common import ResourceReference
