+++ st2common/st2common/models/api/trigger.py
@@ -18,58 +18,10 @@ import uuid
 from st2common.util import isotime
 from st2common.models.api.base import BaseAPI
 from st2common.models.api.tag import TagsHelper
-from st2common.models.db.reactor import SensorTypeDB, TriggerTypeDB, TriggerDB, TriggerInstanceDB
+from st2common.models.db.reactor import TriggerTypeDB, TriggerDB, TriggerInstanceDB
 from st2common.models.system.common import ResourceReference
-from st2common.models.utils import sensor_type_utils
 
 DATE_FORMAT = '%Y-%m-%d %H:%M:%S.%f'
-
-
-class SensorTypeAPI(BaseAPI):
-    model = SensorTypeDB
-    schema = {
-        'type': 'object',
-        'properties': {
-            'id': {
-                'type': 'string',
-                'default': None
-            },
-            'name': {
-                'type': 'string',
-                'required': True
-            },
-            'pack': {
-                'type': 'string'
-            },
-            'description': {
-                'type': 'string'
-            },
-            'artifact_uri': {
-                'type': 'string',
-            },
-            'entry_point': {
-                'type': 'string',
-            },
-            'enabled': {
-                'description': 'Enable or disable the sensor.',
-                'type': 'boolean',
-                'default': True
-            },
-            'trigger_types': {
-                'type': 'array',
-                'default': []
-            },
-            'poll_interval': {
-                'type': 'number'
-            }
-        },
-        'additionalProperties': False
-    }
-
-    @classmethod
-    def to_model(cls, sensor_type):
-        model = sensor_type_utils.to_sensor_db_model(sensor_api_model=sensor_type)
-        return model
 
 
 class TriggerTypeAPI(BaseAPI):
