+++ st2common/tests/unit/test_executions_util.py
@@ -17,7 +17,7 @@ import six
 
 from st2common.constants import action as action_constants
 from st2common.models.api.action import RunnerTypeAPI, ActionAPI, LiveActionAPI
-from st2common.models.api.reactor import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
+from st2common.models.api.trigger import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
 from st2common.models.api.rule import RuleAPI
 from st2common.persistence.action import RunnerType, LiveAction
 from st2common.persistence.execution import ActionExecution
